Hardcode API version and fix bug with query params
This commit removes the required api_version parameter when initializing
a PagerDuty::Connection object. The reason for this change is that since
this gem no longer supports v1 of the REST API and since we don't know
if there will be future versions of the API or what changes will be made
to future versions, it makes sense to hardcode the value for now.

This commit also fixes a bug that occurs when no query parameters are sent
when performing a GET HTTP request. #get incorrectly expected query parameters
to always be sent because #get modifies the contents of the query params hash.

[151816727386211](https://app.asana.com/0/151816727386211/151816727386211)